finish funcs / delete size_t define from libft.h

+++ source/str/ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+unsigned long	ft_strlen(const char *s)
 {
-	size_t	count;
+	unsigned long	count;
 
 	count = 0;
 	while (s[count])
